// src/models/f1/car-frame.ts
export const CAR_FRAME = {
// ...
  parts: {
    "front-wing": {
      FrontWingLivery: {
        min: [-3.012, 0.129, -0.595],
        max: [-2.322, 0.309, 0.595]
      },
      material_0: {
        min: [-2.578, 0.283, -0.878],
        max: [-2.339, 0.33, 0.878]
      },
      FrontEndplateLivery: {
        min: [-2.759, 0.104, -0.882],
        max: [-2.094, 0.394, 0.882]
      },
      Carbon: {
        min: [-2.88, 0.142, -0.004],
        max: [-2.696, 0.183, 0.004]
      }
    },
// ...
  }
} as const;